transfer ViewModels & Models to Domain layer library
