+++ Common/WebStore.Domain/Models/Cart.cs
@@ -1,11 +1,11 @@
 ﻿using System.Collections.Generic;
 using System.Linq;
 
-namespace WebStore.Models
+namespace WebStore.Domain.Models
 {
     public class Cart
     {
         public List<CartItem> Items { get; set; } = new List<CartItem>();
         public int ItemsCount => Items?.Sum(item => item.Quantity) ?? 0;
-}
+    }
 }
