+++ Common/WebStore.Domain/ViewModels/Blog/BlogPostViewModel.cs
@@ -1,17 +1,19 @@
 ﻿using System;
 using System.Collections.Generic;
 
-namespace WebStore.ViewModels.Blog
+namespace WebStore.Domain.ViewModels.Blog
 {
-    public class BlogPostShortInfoViewModel
+    public class BlogPostViewModel
     {
         public int Id { get; set; }
         public UserViewModel Author { get; set; }
         public string Subject { get; set; }
         public DateTime RegistrationTime { get; set; }
         public string MainImageUrl { get; set; }
-        public string ShortDesc { get; set; }
 
+        public List<string> Text { get; set; }
         public List<string> Tags { get; set; }
+
+        public List<CommentViewModel> Comments { get; set; }
     }
 }
