+++ Common/WebStore.Domain/ViewModels/EmployeeViewModel.cs
@@ -1,11 +1,9 @@
 ﻿using System;
 using System.ComponentModel.DataAnnotations;
 
-using WebStore.Domain;
-
-namespace WebStore.ViewModels.Blog
+namespace WebStore.Domain.ViewModels
 {
-    public class UserViewModel
+    public class EmployeeViewModel
     {
         public int Id { get; set; }
 
@@ -24,6 +22,8 @@ namespace WebStore.ViewModels.Blog
         public int Age { get; set; }
 
         public Sex Sex { get; set; }
+        [DataType(DataType.Date)]
+        [DisplayFormat(DataFormatString = "{0:yyyy-MM-dd}", ApplyFormatInEditMode = true)]
         public DateTime Birthday { get; set; }
 
         public string Name { get => $"{Firstname} {Surname}"; }
