+++ Common/WebStore.Domain/ViewModels/Identity/LoginViewModel.cs
@@ -2,7 +2,7 @@
 
 using System.ComponentModel.DataAnnotations;
 
-namespace WebStore.ViewModels.Identity
+namespace WebStore.Domain.ViewModels.Identity
 {
     public class LoginViewModel
     {
