+++ Common/WebStore.Domain/ViewModels/Products/BrandViewModel.cs
@@ -1,6 +1,6 @@
 ﻿using WebStore.Domain.Entities.Base.Interfaces;
 
-namespace WebStore.ViewModels.Products
+namespace WebStore.Domain.ViewModels.Products
 {
     public class BrandViewModel : INamedEntity, IOrderedEntity
     {
