+++ Common/WebStore.Domain/ViewModels/Products/Orders/OrderViewModel.cs
@@ -1,6 +1,6 @@
 ﻿using System.ComponentModel.DataAnnotations;
 
-namespace WebStore.ViewModels.Products.Orders
+namespace WebStore.Domain.ViewModels.Products.Orders
 {
     public class OrderViewModel
     {
