+++ UI/WebStore/Areas/Admin/Views/_ViewImports.cshtml
@@ -1,5 +1,5 @@
 ﻿@using WebStore
-@using WebStore.ViewModels
+@using WebStore.Domain.ViewModels
 
 @addTagHelper *, Microsoft.AspNetCore.Mvc.TagHelpers
 @addTagHelper *, WebStore
