+++ UI/WebStore/Components/BrandsViewComponent.cs
@@ -5,8 +5,8 @@ using Microsoft.AspNetCore.Mvc;
 using System.Collections.Generic;
 using System.Linq;
 
+using WebStore.Domain.ViewModels.Products;
 using WebStore.Infrastructure.Interfaces;
-using WebStore.ViewModels.Products;
 
 namespace WebStore.Components
 {
