+++ UI/WebStore/Controllers/BlogController.cs
@@ -4,9 +4,9 @@ using System.Collections.Generic;
 using System.Linq;
 
 using WebStore.Domain.Entities;
+using WebStore.Domain.ViewModels.Blog;
 using WebStore.Infrastructure.Interfaces;
 using WebStore.Infrastructure.Services.InMemory;
-using WebStore.ViewModels.Blog;
 
 namespace WebStore.Controllers
 {
