+++ UI/WebStore/Infrastructure/Profiles/EmployeeProfile.cs
@@ -1,7 +1,7 @@
 ﻿using AutoMapper;
 
 using WebStore.Domain.Entities;
-using WebStore.ViewModels;
+using WebStore.Domain.ViewModels;
 
 namespace WebStore.Infrastructure.Profiles
 {
