+++ UI/WebStore/Views/_ViewImports.cshtml
@@ -1,9 +1,9 @@
 ﻿@using WebStore
-@using WebStore.ViewModels
-@using WebStore.ViewModels.Products
-@using WebStore.ViewModels.Products.Orders
-@using WebStore.ViewModels.Blog
-@using WebStore.ViewModels.Identity
+@using WebStore.Domain.ViewModels
+@using WebStore.Domain.ViewModels.Products
+@using WebStore.Domain.ViewModels.Products.Orders
+@using WebStore.Domain.ViewModels.Blog
+@using WebStore.Domain.ViewModels.Identity
 
 @addTagHelper *, Microsoft.AspNetCore.Mvc.TagHelpers
 @addTagHelper *, WebStore
